re-implementing isis cubefile
